% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj-model-utils.R
\name{add_features_r2}
\alias{add_features_r2}
\title{Add Features R^2}
\usage{
add_features_r2(traj_model, sample_frac = 0.1, seed = 60427)
}
\arguments{
\item{traj_model}{The trajectory model object.}

\item{sample_frac}{The fraction of samples to use for computing the r2 without each model. When NULL, all samples are used.}
}
\value{
The trajectory model object with the added R-squared values.
}
\description{
This function adds the added R-squared values of each feature to the trajectory model.
}
