% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj-model-utils.R
\name{adjust_energies}
\alias{adjust_energies}
\title{Adjust Energy Values in a Trajectory Model}
\usage{
adjust_energies(
  traj_model,
  q = c(1, 0.999, 0.995, 0.99, 0.98, 0.97, 0.96, 0.95),
  min_energy = c(-20, -15, -13, -10, -7, -5),
  diff_thresh = 0.01,
  sequences = NULL,
  norm_sequences = NULL,
  relearn = TRUE,
  lambda = NULL,
  use_additional_features = TRUE,
  use_motifs = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{traj_model}{The trajectory model object containing the motif models and peak intervals.}

\item{q}{A numeric vector of quantile values to test for normalization. Default values range from 1 to 0.95.}

\item{min_energy}{A numeric vector of minimum energy values to test. Default values range from -20 to -5.}

\item{diff_thresh}{The threshold for the difference in correlation between the current and best parameters. Default is 0.01.}

\item{sequences}{Optional character vector of DNA sequences corresponding to the peak intervals.
If NULL, sequences will be extracted from the trajectory model.}

\item{norm_sequences}{Optional character vector of DNA sequences for normalization.
If NULL, sequences will be extracted from the trajectory model.}

\item{relearn}{Logical indicating whether to relearn the model after adjusting energies. Default is TRUE.}

\item{lambda}{The lambda value to use for relearning. If NULL, the lambda value from the trajectory model is used.}

\item{use_additional_features}{A logical value indicating whether to use additional features. Default is TRUE.}

\item{use_motifs}{A logical value indicating whether to use motif models. Default is TRUE.}

\item{verbose}{Logical indicating whether to display additional information.}
}
\value{
The updated trajectory model object with optimized energy values for each motif.
}
\description{
This function optimizes the energy normalization parameters for each motif in a trajectory model
by testing different combinations of quantile thresholds and minimum energy values. For each motif,
it selects the parameters that maximize the absolute correlation with the differential score.
}
\details{
The function works by:
\enumerate{
\item Extracting DNA sequences from the peak intervals if not provided
\item Computing motif energies for all sequences
\item Testing different combinations of normalization parameters
\item For each motif, selecting the parameters that maximize correlation with the differential score
\item Applying the selected parameters to normalize the energies
}
}
\examples{
\dontrun{
traj_model <- adjust_energies(traj_model)

# Adjust energies with custom parameters
traj_model <- adjust_energies(
    traj_model,
    q = c(1, 0.99, 0.95),
    min_energy = c(-15, -10, -5)
)
}

}
