% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utils.R
\name{bin_atac_on_traj}
\alias{bin_atac_on_traj}
\title{Bin ATAC data based on trajectory metacells}
\usage{
bin_atac_on_traj(
  mc_atac,
  traj_metacells,
  const_peak_score_all = NULL,
  traj_metacells_order = 1:length(traj_metacells),
  bins = 2,
  norm_quant = 1
)
}
\arguments{
\item{mc_atac}{The ATAC matrix.}

\item{traj_metacells}{A vector of trajectory metacells.}

\item{const_peak_score_all}{A vector of constant peak scores for all metacells (optional).}

\item{traj_metacells_order}{The order of trajectory metacells (optional).}

\item{bins}{The number of bins to use for binning (default is 2).}

\item{norm_quant}{The quantile to use for normalization (default is 1).}
}
\value{
A list containing the binned ATAC data (ATAC_bin), a mapping of metacells to quantiles (mcs2quant)
and the normalized ATAC data (ATAC_o_s_nrm).
}
\description{
This function bins ATAC data based on trajectory metacells. It takes as input an ATAC matrix, a vector of trajectory metacells, and optional parameters for normalization and binning. It returns a list containing the binned ATAC data, a mapping of metacells to quantiles, and the normalized ATAC data.
}
