% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-energies.R
\name{compute_motif_energies}
\alias{compute_motif_energies}
\title{Compute Motif Energies}
\usage{
compute_motif_energies(
  peak_intervals = NULL,
  peaks = NULL,
  db = iceqream::motif_db,
  normalization_intervals = peak_intervals,
  prior = 0.01,
  normalize = TRUE,
  energy_norm_quantile = 1,
  norm_energy_max = 10,
  min_energy = -7,
  db_quantiles = NULL
)
}
\arguments{
\item{peak_intervals}{A data frame containing the peak intervals with columns: chrom, start, and end. If an additional column 'peak_name' is present, it will be used as row names in the output matrix.}

\item{peaks}{An alias for \code{peak_intervals}.}

\item{db}{A motif database to use for extracting motif energies. Default is \code{iceqream::motif_db}.}

\item{normalization_intervals}{A data frame containing intervals for normalization. Default is \code{peak_intervals}.}

\item{prior}{a prior probability for each nucleotide.}

\item{normalize}{A logical value indicating whether to normalize the motif energies. Default is TRUE.}

\item{energy_norm_quantile}{A numeric value for the quantile used in normalization. Default is 1.}

\item{norm_energy_max}{The maximum value to which the normalized energy values are scaled. Default is 10.}

\item{min_energy}{The minimum energy value to be assigned after normalization. Default is -7.}

\item{db_quantiles}{A matrix of motif energies quantiles to use for normalization. Rows are motifs and columns are quantiles. If not NULL, this would be used instead of computing the quantiles from the normalization intervals. A precomputed matrix for mouse gastrulation can be found in \code{iceqream::mouse_db_quantiles}.}
}
\value{
A matrix of motif energies.
}
\description{
This function computes motif energies for given peak intervals using a specified motif database.
Optionally, it can normalize the computed motif energies using additional normalization intervals.
}
