% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq-plot.R
\name{compute_tracks_q}
\alias{compute_tracks_q}
\title{Compute tracks quantiles}
\usage{
compute_tracks_q(
  atac_names,
  atac_tracks,
  iterator = 20,
  norm_q = 0.995,
  norm_intervals = gintervals.all(),
  tn5bias_track = "tn5bias",
  normalize_tn5bias = TRUE,
  sample_n = 10000
)
}
\arguments{
\item{atac_names}{A character vector specifying the names of the ATAC-seq tracks.}

\item{atac_tracks}{A list of numeric vectors representing the ATAC-seq tracks.}

\item{iterator}{An iterator object used for computing quantiles.}

\item{norm_q}{A numeric vector specifying the quantiles to compute for each track. If a single value, the same quantile is used for all tracks.}

\item{norm_intervals}{A numeric vector specifying the intervals for computing quantiles.}

\item{tn5bias_track}{A numeric vector representing the TN5 bias track.}

\item{normalize_tn5bias}{Logical, whether to normalize the TN5 bias track.}

\item{sample_n}{A numeric value specifying the maximum number of data points to use for computing quantiles.}
}
\value{
A data frame with columns "type" and "q", where "type" represents the name of the ATAC-seq track and "q" represents the computed quantiles.
}
\description{
This function computes the quantiles for a set of ATAC-seq tracks.
}
