% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{compute_traj_model_spatial_freq}
\alias{compute_traj_model_spatial_freq}
\title{Compute spatial frequency of motifs in trajectory model}
\usage{
compute_traj_model_spatial_freq(
  traj_model,
  size,
  pwm_threshold = 7,
  top_q = 0.1,
  bottom_q = 0.1,
  atac_track = NULL,
  parallel = TRUE
)
}
\arguments{
\item{traj_model}{A trajectory model object}

\item{size}{The size of the region to compute spatial frequency for}

\item{pwm_threshold}{The threshold for the PWM score}

\item{top_q}{The proportion of top peaks to select}

\item{bottom_q}{The proportion of bottom peaks to select}
}
\value{
A data frame with the spatial frequency of each motif
}
\description{
This function computes the spatial frequency of motifs in a trajectory model, using the top and bottom 10\% of peaks based on diff_score.
}
