% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQSeqFeature.R
\name{create_cg_content_feature}
\alias{create_cg_content_feature}
\title{Create CG Content Feature}
\usage{
create_cg_content_feature(traj_model, quantile = 0.99)
}
\arguments{
\item{traj_model}{A trajectory model object.}

\item{quantile}{The quantile to use for normalization (default is 0.99).}
}
\value{
An IQSeqFeature object for CG content.
}
\description{
This function creates an IQSeqFeature that calculates -log2(1-CG_fraction) as a measure
of CG content in DNA sequences. The feature is normalized to be between 0-10.
}
