% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{create_default_additional_features}
\alias{create_default_additional_features}
\title{Create Default Additional Features}
\usage{
create_default_additional_features(peaks, tracks, window_size = 20000)
}
\arguments{
\item{peaks}{Intervals set of peaks}

\item{tracks}{Tracks containing ATAC signal}

\item{window_size}{An integer specifying the size of the window for computing regional ATAC signal. Default is 20,000 bp.}
}
\value{
A data frame with additional features: sequence features and regional ATAC signal. The features are normalized to the range 0-10.
}
\description{
This function generates additional features for a given set of peaks. The features include sequence features and regional ATAC signal using a punctured window.
}
