% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQSeqFeature.R
\name{create_dinuc_feature_group}
\alias{create_dinuc_feature_group}
\title{Create Dinucleotide Feature Group}
\usage{
create_dinuc_feature_group(
  traj_model,
  quantile = 0.99,
  size = traj_model@params$peaks_size,
  dinucleotides = c("AA", "AC", "AG", "AT", "CA", "CC", "CG", "CT", "GA", "GC", "GG",
    "GT", "TA", "TC", "TG", "TT")
)
}
\arguments{
\item{traj_model}{A trajectory model object.}

\item{quantile}{The quantile to use for normalization (default is 0.99).}

\item{size}{The size of the sequences to use for feature computation.}

\item{dinucleotides}{A vector of dinucleotides to create features for (default is all possible dinucleotides).}
}
\value{
An IQFeatureGroup object containing all dinucleotide features.
}
\description{
This function creates an IQFeatureGroup object for all dinucleotide features.
}
