% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQSeqFeature.R
\name{create_dinuc_features}
\alias{create_dinuc_features}
\title{Create Dinucleotide Features}
\usage{
create_dinuc_features(
  traj_model,
  quantile = 0.99,
  size = traj_model@params$peaks_size,
  dinucleotides = c("AA", "AC", "AG", "AT", "CA", "CC", "CG", "CT", "GA", "GC", "GG",
    "GT", "TA", "TC", "TG", "TT"),
  feat_names = dinucleotides
)
}
\arguments{
\item{traj_model}{A trajectory model object.}

\item{quantile}{The quantile to use for normalization (default is 0.99).}

\item{size}{The size of the sequences to use for feature computation.}

\item{dinucleotides}{A vector of dinucleotides to create features for (default is all possible dinucleotides).}

\item{feat_names}{A vector of feature names to use for each dinucleotide (default is the dinucleotide itself).}
}
\value{
A list of IQSeqFeature objects, one for each dinucleotide.
}
\description{
This function creates IQSeqFeature objects for all possible dinucleotides.
}
