% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQModel.R
\name{create_iq_model}
\alias{create_iq_model}
\title{Create an IQmodel from a trajectory model}
\usage{
create_iq_model(traj_model, func = "logSumExp")
}
\arguments{
\item{traj_model}{A trajectory model object}

\item{func}{The function to use for computing energies (default is "logSumExp")}
}
\value{
An IQmodel object
}
\description{
This function creates an IQmodel object from a trajectory model.
}
