% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQSeqFeature.R
\name{create_iq_seq_feature}
\alias{create_iq_seq_feature}
\title{Create an IQSeqFeature from a trajectory model}
\usage{
create_iq_seq_feature(traj_model, feature_name, compute_func, quantile = 0.99)
}
\arguments{
\item{traj_model}{A trajectory model object.}

\item{feature_name}{The name of the feature to create.}

\item{compute_func}{The function to compute the feature.}

\item{quantile}{The quantile to use for normalization (default is 0.99).}
}
\value{
An IQSeqFeature object.
}
\description{
Create an IQSeqFeature from a trajectory model
}
