% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-utils.R
\name{create_logist_features}
\alias{create_logist_features}
\title{Create Logistic Features}
\usage{
create_logist_features(features)
}
\arguments{
\item{features}{A matrix or dataframe where each column is a feature to be transformed.}
}
\value{
A matrix containing the transformed features with columns named according to
the transformation applied (i.e., "_early", "_linear", or "_late").
}
\description{
This function takes a matrix or dataframe of features, removes columns that are entirely NA,
and then applies three logistic transformations to each column. Each transformed set of features
is appended with suffixes "_early", "_linear", or "_late" to differentiate between them.
The resulting matrix combines all transformed features.
}
\examples{
# Create a sample matrix
sample_features <- matrix(rnorm(100), ncol = 5)
transformed_features <- create_logist_features(sample_features)
head(transformed_features)

}
\seealso{
\code{\link{logist}} for the logistic transformation function.
}
