% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill-motifs.R
\name{distill_traj_model}
\alias{distill_traj_model}
\title{Distill trajectory model}
\usage{
distill_traj_model(
  traj_model,
  max_motif_num,
  min_diff = 0.1,
  intra_cor_thresh = 0.6,
  use_non_linear = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{traj_model}{A trajectory model object}

\item{max_motif_num}{The maximum number of motifs to select}

\item{min_diff}{The minimum difference of scores to use for distillation}

\item{intra_cor_thresh}{The threshold for the average intra-cluster correlation to split clusters. If NULL, no splitting is done.}

\item{use_non_linear}{Whether to use non-linear features. If FALSE, a glm model without non-linear features is used.}

\item{parallel}{Whether to use parallel processing}
}
\value{
A distilled trajectory model object with \code{max_motif_num} models + additional features.
}
\description{
This function takes a trajectory model and reduces the number of motifs using 'prego'.
}
