% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill-multi-traj.R
\name{distill_traj_model_multi}
\alias{distill_traj_model_multi}
\title{Distills multiple trajectory models}
\usage{
distill_traj_model_multi(
  traj_models,
  max_motif_num = NULL,
  min_diff = 0.1,
  intra_cor_thresh = 0.6,
  distill_single = FALSE,
  use_all_motifs = FALSE,
  bits_threshold = 1.75,
  r2_threshold = NULL,
  seed = 60427,
  parallel = TRUE,
  cluster_report_dir = NULL
)
}
\arguments{
\item{traj_models}{A list of trajectory models.}

\item{max_motif_num}{The maximum number of motifs to be identified. If NULL - the number would be set to the maximum number of motifs in the input models.}

\item{min_diff}{The minimum difference of scores to use for distillation}

\item{intra_cor_thresh}{The threshold for the average intra-cluster correlation to split clusters. If NULL, no splitting is done.}

\item{distill_single}{Logical indicating whether to distill clusters with a single motif.}

\item{use_all_motifs}{Logical indicating whether to use all motifs in the resulting models. If FALSE, only motifs from clusters which had a motif from the original model are used.}

\item{bits_threshold}{minimal sum of bits for a feature to be included in the model.}

\item{r2_threshold}{minimal R^2 for a feature to be included in the model.}

\item{seed}{The random seed for reproducibility. Defaults to 60427.}

\item{parallel}{Whether to use parallel processing}

\item{cluster_report_dir}{The directory to store cluster reports. If not NULL, a png would be created for each cluster.}
}
\value{
A distilled trajectory model.
}
\description{
This function takes a list of trajectory models and performs a distillation process to create a single set of motifs.
}
