% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eflow.R
\docType{class}
\name{Eflow-class}
\alias{Eflow-class}
\alias{Eflow}
\title{EFlow object}
\description{
Eflow object is ...
}
\section{Slots}{

\describe{
\item{\code{id}}{an identifier for the object, e.g. "gastrulation".}

\item{\code{description}}{description of the object, e.g. "Gastrulation flow"}
}}

