% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhflow-package.R
\docType{package}
\name{enhflow-package}
\alias{enhflow}
\alias{enhflow-package}
\title{enhflow: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

Authors:
\itemize{
  \item Akhiad Bercovich \email{akhiad.bercovich@weizmann.ac.il}
  \item Amos Tanay \email{amos.tanay@weizmann.ac.il}
}

}
\keyword{internal}
