% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill-multi-traj.R
\name{filter_multi_traj_model}
\alias{filter_multi_traj_model}
\title{Filter multi-trajectory model}
\usage{
filter_multi_traj_model(
  multi_traj,
  r2_threshold = 0.0005,
  bits_threshold = 1.75,
  sample_frac = 0.1,
  filter_full = TRUE
)
}
\arguments{
\item{r2_threshold}{minimal R^2 for a feature to be included in the model.}

\item{bits_threshold}{minimal sum of bits for a feature to be included in the model.}

\item{sample_frac}{The fraction of samples to use for computing the r2 without each model. When NULL, all samples are used.}

\item{filter_full}{A logical value indicating whether to filter the full models (\code{@models_full}) or the reduced models (\code{@models}). Default is TRUE.}
}
\value{
The filtered multi-trajectory model.
}
\description{
Filter multi-trajectory model
}
