% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill-multi-traj.R
\name{filter_multi_traj_model_by_beta}
\alias{filter_multi_traj_model_by_beta}
\title{Filter multi-trajectory model by beta}
\usage{
filter_multi_traj_model_by_beta(
  multi_traj,
  beta_threshold = 0.005,
  filter_full = TRUE,
  unify = TRUE
)
}
\arguments{
\item{multi_traj}{A TrajectoryModelMulti model object.}

\item{beta_threshold}{The threshold for the absolute value of the beta coefficient. Default is 0.005.}

\item{filter_full}{A logical value indicating whether to filter the full models (\code{@models_full}) or the reduced models (\code{@models}). Default is TRUE.}

\item{unify}{Whether to use all the motifs in each trajectory after filtering}
}
\value{
The filtered TrajectoryModelMulti object
}
\description{
Filter multi-trajectory model by beta
}
