% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-model.R
\name{filter_traj_model}
\alias{filter_traj_model}
\title{Filter a trajectory model using punctuated regression.}
\usage{
filter_traj_model(traj_model, r2_threshold = 0.0005)
}
\arguments{
\item{traj_model}{An instance of \code{TrajectoryModel}.}

\item{r2_threshold}{minimal R^2 for a feature to be included in the model.}
}
\value{
An instance of \code{TrajectoryModel} with the filtered model.
}
\description{
Run a regression without each feature and filter features that do not improve the model more
than \code{r2_threshold}.
}
