% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj-model-utils.R
\name{filter_traj_model_intervals}
\alias{filter_traj_model_intervals}
\title{Filter trajectory model intervals}
\usage{
filter_traj_model_intervals(traj_model, idxs)
}
\arguments{
\item{traj_model}{The trajectory model object to filter.}

\item{idxs}{The indices of the intervals to select.}
}
\value{
The filtered trajectory model object.
}
\description{
This function filters a trajectory model by selecting specific intervals based on their indices.
}
