% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prego.R
\name{infer_traj_prego}
\alias{infer_traj_prego}
\title{Infer 'prego' models for ATAC difference of a trajectory}
\usage{
infer_traj_prego(
  peak_intervals,
  atac_diff,
  n_motifs,
  min_diff = 0.2,
  energy_norm_quantile = 0.99,
  sample_fraction = 0.1,
  sequences = NULL,
  seed = NULL
)
}
\arguments{
\item{peak_intervals}{A data frame, indicating the genomic positions ('chrom', 'start', 'end') of each peak}

\item{atac_diff}{A numeric vector, indicating the ATAC difference of each peak}

\item{n_motifs}{Number of motifs to infer}

\item{min_diff}{Minimum ATAC difference to include a peak in the training}

\item{energy_norm_quantile}{quantile of the energy used for normalization. Default: 0.99}

\item{sample_fraction}{Fraction of peaks to sample for training. Default: 0.1}

\item{sequences}{A character vector of sequences to infer the motifs on. If NULL, the sequences of the peaks are used.}

\item{seed}{Random seed}
}
\description{
Infer 'prego' models for ATAC difference of a trajectory
}
