% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill-multi-traj.R
\name{infer_trajectory_motifs_multi}
\alias{infer_trajectory_motifs_multi}
\title{Infer trajectory motifs for a multi-trajectory model}
\usage{
infer_trajectory_motifs_multi(
  traj_multi,
  peak_intervals,
  atac_scores = NULL,
  bin_start = 1,
  bin_end = purrr::map(atac_scores, ncol),
  additional_features = NULL
)
}
\arguments{
\item{traj_multi}{A TrajectoryModelMulti object.}

\item{peak_intervals}{A data frame, indicating the genomic positions ('chrom', 'start', 'end') of each peak.}

\item{atac_scores}{A list of data frames with ATAC-seq scores for each individual trajectory.}

\item{bin_start}{the start of the trajectory. Default: 1}

\item{bin_end}{the end of the trajectory. Default: the last bin (only used when atac_scores is provided)}

\item{additional_features}{A list of data frames with additional features for each individual trajectory.}
}
\value{
A TrajectoryModelMulti object with the result of \code{infer_trajectory_motifs} for each individual trajectory.
}
\description{
This function infers trajectory motifs for each individual trajectory in a multi-trajectory model.
}
