% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eflow.R
\name{init}
\alias{init}
\title{Initialize an EFlow object}
\usage{
init(mc2_file, md, id = NULL, description = NULL)
}
\arguments{
\item{mc2_file}{Path to metacell2 (metacells python package) Anndata file}

\item{md}{A data frame with a column named "metacell" with the metacell IDs, and additional columns with metadata, or a file
which contains such data frame.}

\item{id}{an identifier for the object, e.g. "gastrulation".}

\item{description}{description of the object, e.g. "Gastrulation flow"}
}
\value{
an Eflow object
}
\description{
Initialize an EFlow object
}
