% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQFeatureGroup.R
\name{initialize,IQFeatureGroup-method}
\alias{initialize,IQFeatureGroup-method}
\title{Initialize an IQFeatureGroup object}
\usage{
\S4method{initialize}{IQFeatureGroup}(.Object, features, compute_func, size)
}
\arguments{
\item{.Object}{The IQFeatureGroup object to initialize.}

\item{features}{A list of IQSeqFeature objects.}

\item{compute_func}{A function to compute all features in the group at once.}

\item{size}{The size of the sequences to use for feature computation.}
}
\value{
An initialized IQFeatureGroup object.
}
\description{
Initialize an IQFeatureGroup object
}
