% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQSeqFeature.R
\name{initialize,IQSeqFeature-method}
\alias{initialize,IQSeqFeature-method}
\title{Initialize an IQSeqFeature object}
\usage{
\S4method{initialize}{IQSeqFeature}(
  .Object,
  name,
  coefs,
  compute_func,
  min_value,
  max_value,
  quantile = 0.95,
  size
)
}
\arguments{
\item{.Object}{The IQSeqFeature object to initialize.}

\item{name}{The name of the feature.}

\item{coefs}{The coefficients of the feature.}

\item{compute_func}{The function to compute the feature.}

\item{min_value}{The minimum value for normalization.}

\item{max_value}{The maximum value for normalization.}

\item{quantile}{The quantile to use for normalization.}

\item{size}{The size of the sequences to use for feature computation.}
}
\value{
An initialized IQSeqFeature object.
}
\description{
Initialize an IQSeqFeature object
}
