% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQFeature.R
\name{iq_feature.compute}
\alias{iq_feature.compute}
\title{Compute the IQ feature}
\usage{
iq_feature.compute(iq, values)
}
\arguments{
\item{iq}{An IQFeature object.}

\item{values}{A vector of values to compute the IQ feature.}
}
\value{
A vector of computed IQ feature responses.
}
\description{
This function computes the IQ feature response for a given set of values.
}
