% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQFeatureGroup.R
\name{iq_feature_group.compute}
\alias{iq_feature_group.compute}
\title{Compute features for an IQFeatureGroup}
\usage{
iq_feature_group.compute(group, sequences = NULL, intervals = NULL)
}
\arguments{
\item{group}{An IQFeatureGroup object.}

\item{sequences}{A vector of DNA sequences. Optional if intervals are provided.}

\item{intervals}{A data frame of genomic intervals with columns 'chrom', 'start', and 'end'. Optional if sequences are provided.}
}
\value{
A matrix of computed and normalized feature values.
}
\description{
This function computes all features in the group for a given set of sequences or intervals.
}
