% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQFeature.R
\name{iq_feature_list.compute}
\alias{iq_feature_list.compute}
\title{Compute IQ feature list}
\usage{
iq_feature_list.compute(iq_list, mat)
}
\arguments{
\item{iq_list}{A list of IQFeature objects}

\item{mat}{The input matrix}
}
\value{
A matrix containing the computed IQ feature responses.
}
\description{
This function computes the IQ response for a given list of IQ features and a matrix of values.
}
