% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQInteraction.R
\name{iq_interaction.compute}
\alias{iq_interaction.compute}
\title{Compute an IQInteraction}
\usage{
iq_interaction.compute(iq, term1_values, term2_values)
}
\arguments{
\item{iq}{An IQInteraction object.}

\item{term1_values}{A vector of values for the first term.}

\item{term2_values}{A vector of values for the second term.}
}
\value{
A vector of computed and normalized interaction values.
}
\description{
This function computes the value of an IQ interaction feature given the values of its terms.
}
