% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQInteraction.R
\name{iq_model.compute_interactions}
\alias{iq_model.compute_interactions}
\title{Compute all interaction features}
\usage{
iq_model.compute_interactions(model, feature_data)
}
\arguments{
\item{model}{An IQmodel object}

\item{feature_data}{A data frame containing feature values needed for interactions}
}
\value{
A data frame containing computed interaction values
}
\description{
This function computes all interaction features for a given dataset.
}
