% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQInteraction.R
\name{iq_model.feature_names_by_type}
\alias{iq_model.feature_names_by_type}
\title{Get all feature names from IQ features with categorization by type}
\usage{
iq_model.feature_names_by_type(iq_features, include_interaction_terms = FALSE)
}
\arguments{
\item{iq_features}{List of IQ features}

\item{include_interaction_terms}{Logical indicating whether to include interaction terms}
}
\value{
A list with named vectors for different feature types
}
\description{
Get all feature names from IQ features with categorization by type
}
