% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQInteraction.R
\name{iq_model.get_interactions}
\alias{iq_model.get_interactions}
\title{Get interaction features from an IQmodel}
\usage{
iq_model.get_interactions(model)
}
\arguments{
\item{model}{An IQmodel object}
}
\value{
A list of IQInteraction objects
}
\description{
This function extracts all interaction features from an IQmodel.
}
