% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQInteraction.R
\name{iq_model.get_regular_features}
\alias{iq_model.get_regular_features}
\title{Get regular features from an IQmodel}
\usage{
iq_model.get_regular_features(model)
}
\arguments{
\item{model}{An IQmodel object}
}
\value{
A list of IQFeature objects (excluding interactions)
}
\description{
This function extracts all non-interaction features from an IQmodel.
}
