% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQInteraction.R
\name{iq_model.has_interactions}
\alias{iq_model.has_interactions}
\title{Check if an IQmodel has interaction features}
\usage{
iq_model.has_interactions(model)
}
\arguments{
\item{model}{An IQmodel object}
}
\value{
Logical indicating whether the model has interactions
}
\description{
This function checks if an IQmodel contains any interaction features.
}
