% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQSeqFeature.R
\name{iq_seq_feature.compute}
\alias{iq_seq_feature.compute}
\title{Compute the IQSeqFeature}
\usage{
iq_seq_feature.compute(iq, sequences)
}
\arguments{
\item{iq}{An IQSeqFeature object.}

\item{sequences}{A vector of DNA sequences.}
}
\value{
A vector of computed and normalized IQ sequence feature values.
}
\description{
This function computes the IQ sequence feature for a given set of sequences.
}
