% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQFeature.R
\docType{class}
\name{IQFeature-class}
\alias{IQFeature-class}
\alias{IQFeature}
\title{IQFeature class}
\description{
This class represents an IQ feature
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the IQ feature (character).}

\item{\code{coefs}}{The coefficients of the IQ feature (numeric).}
}}

