% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQFeatureGroup.R
\docType{class}
\name{IQFeatureGroup-class}
\alias{IQFeatureGroup-class}
\alias{IQFeatureGroup}
\title{IQFeatureGroup class}
\description{
This class represents a group of related IQ features, such as all dinucleotide features.
}
\section{Slots}{

\describe{
\item{\code{features}}{A list of IQSeqFeature objects.}

\item{\code{compute_func}}{A function to compute all features in the group at once.}

\item{\code{size}}{The size of the sequences to use for feature computation (numeric).}
}}

