% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQInteraction.R
\docType{class}
\name{IQInteraction-class}
\alias{IQInteraction-class}
\alias{IQInteraction}
\alias{show,IQInteraction-method}
\title{IQInteraction class}
\usage{
\S4method{show}{IQInteraction}(object)
}
\arguments{
\item{object}{An IQInteraction object}
}
\description{
This class represents an interaction feature for IQ models, inheriting from IQFeature.

This method defines how an IQInteraction object should be displayed.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the IQ interaction feature (character).}

\item{\code{coefs}}{The coefficients of the IQ interaction feature (numeric).}

\item{\code{term1}}{The name of the first term in the interaction (character).}

\item{\code{term2}}{The name of the second term in the interaction (character).}

\item{\code{term1_type}}{The type of the first term (motif, additional, etc.).}

\item{\code{term2_type}}{The type of the second term (motif, additional, etc.).}
}}

