% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQModel.R
\docType{class}
\name{IQmodel-class}
\alias{IQmodel-class}
\alias{IQmodel}
\title{IQmodel class}
\description{
This class represents an IQ model including IQ features and PBMs,
along with necessary model parameters.
}
\section{Slots}{

\describe{
\item{\code{iq_features}}{A list of IQFeature objects.}

\item{\code{pbms}}{A list of PBM objects.}

\item{\code{intercept}}{The intercept term from the model.}

\item{\code{func}}{The function used for computing energies (e.g., "logSumExp" or "max").}

\item{\code{lambda}}{The regularization parameter.}

\item{\code{alpha}}{The elastic net mixing parameter.}

\item{\code{min_pred}}{Minimum value of training predictions before normalization.}

\item{\code{max_pred}}{Maximum value of training predictions before normalization.}

\item{\code{norm_factors}}{Normalization factors for scaling predictions.}
}}

