% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQSeqFeature.R
\docType{class}
\name{IQSeqFeature-class}
\alias{IQSeqFeature-class}
\alias{IQSeqFeature}
\title{IQSeqFeature class}
\description{
This class represents a sequence feature for IQ models, inheriting from IQFeature.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the IQ sequence feature (character).}

\item{\code{coefs}}{The coefficients of the IQ sequence feature (numeric).}

\item{\code{compute_func}}{The function to compute the feature (function).}

\item{\code{min_value}}{The minimum value for normalization (numeric).}

\item{\code{max_value}}{The maximum value for normalization (numeric).}

\item{\code{quantile}}{The quantile to use for normalization (numeric).}

\item{\code{size}}{The size of the sequences to use for feature computation (numeric).}
}}

