% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prego.R
\name{learn_traj_prego}
\alias{learn_traj_prego}
\title{Learn 'prego' models for ATAC difference of a trajectory}
\usage{
learn_traj_prego(
  peak_intervals = NULL,
  peaks = NULL,
  atac_diff,
  n_motifs,
  min_diff = 0.2,
  energy_norm_quantile = 1,
  norm_energy_max = 10,
  min_energy = -7,
  sample_for_kmers = FALSE,
  sample_fraction = 0.1,
  sequences = NULL,
  seed = NULL,
  peaks_size = 300,
  additional_features = NULL,
  norm_intervals = peak_intervals,
  ...
)
}
\arguments{
\item{peak_intervals}{A data frame, indicating the genomic positions ('chrom', 'start', 'end') of each peak, with an additional column named "const" indicating whether the peak is constitutive and therefore shouldn't be used in the regression. Optionally, a column named "cluster" can be added with indication of the cluster of each peak.}

\item{peaks}{An alias for \code{peak_intervals}.}

\item{atac_diff}{A numeric vector, indicating the ATAC difference of each peak}

\item{n_motifs}{Number of motifs to learn. Should be at least 2}

\item{min_diff}{Minimum ATAC difference to include a peak in the training}

\item{energy_norm_quantile}{quantile of the energy used for normalization. Default: 1}

\item{norm_energy_max}{maximum value of the normalized energy. Default: 10}

\item{min_energy}{Minimum energy value after normalization (default: -7)}

\item{sample_for_kmers}{Whether to sample kmers for training. Default: TRUE}

\item{sample_fraction}{Fraction of peaks to sample for training. Default: 0.1#'}

\item{sequences}{A character vector of sequences to learn the motifs on. If NULL, the sequences of the peaks are used.}

\item{seed}{Random seed}

\item{peaks_size}{size of the peaks to extract sequences from. Default: 300bp}

\item{additional_features}{A matrix of additional features to filter out before learning the motifs (e.g. CpG content, dinucleotide content, etc.)}

\item{norm_intervals}{A data frame, indicating the genomic positions ('chrom', 'start', 'end') of peaks used for energy normalization. If NULL, the function will use \code{peak_intervals} for normalization.}

\item{...}{Additional arguments to be passed to \code{prego::regress_pwm}}
}
\description{
Learn 'prego' models for ATAC difference of a trajectory
}
