% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-utils.R
\name{logist}
\alias{logist}
\title{Logistic Function}
\usage{
logist(x, x_0 = 0, L = 1, k = 1)
}
\arguments{
\item{x}{Numeric vector, the values at which the logistic function will be evaluated.}

\item{x_0}{Numeric, the x-value of the sigmoid's midpoint. Default is 0.}

\item{L}{Numeric, the maximum value of the sigmoid. Default is 1.}

\item{k}{Numeric, the steepness or slope of the sigmoid. Default is 1.}
}
\value{
A numeric vector of logistic function values.
}
\description{
Calculates the logistic function value given parameters and input.
}
\examples{
x_vals <- seq(0, 10, by = 0.1)
logist(x_vals, x_0 = 0, L = 2, k = 0.5)

}
