% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj-model-utils.R
\name{match_traj_model_motif_names}
\alias{match_traj_model_motif_names}
\title{Match trajectory model motif names}
\usage{
match_traj_model_motif_names(
  traj_model,
  dataset = prego::all_motif_datasets() \%>\% filter(dataset == "HOMER")
)
}
\arguments{
\item{traj_model}{The trajectory model object.}

\item{dataset}{The dataset to match the motif names with. Default is the "HOMER" dataset.}
}
\value{
A named character vector mapping the motif names in the trajectory model to the matched motif names in the dataset.
}
\description{
This function matches the motif names in a trajectory model with a given dataset.
This is used in order to give more 'friendly' names to the motif models.
The default dataset is "HOMER".
Note that the run might be slow.
}
