% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_db_scenic_clusters}
\alias{motif_db_scenic_clusters}
\title{SCENIC Motif Clusters}
\format{
The same as \code{motif_db} but filtered to only include SCENIC motifs.
}
\source{
\url{https://doi.org/10.1038/nmeth.4463}
}
\usage{
motif_db_scenic_clusters
}
\description{
A data frame containing a representative motif from each SCENIC cluster.
}
\examples{
data(motif_db_scenic_clusters)
head(motif_db_scenic_clusters)
}
\seealso{
\code{\link{motif_db}}
}
\keyword{datasets}
