% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-utils.R
\name{norm0q}
\alias{norm0q}
\title{Normalize a vector to a specified quantile}
\usage{
norm0q(x, quant = 0.99)
}
\arguments{
\item{x}{A numeric vector to be normalized}

\item{quant}{The quantile to normalize to (default is 0.99)}
}
\value{
A normalized vector
}
\description{
This function takes a vector \code{x} and normalizes it to a specified quantile.
The normalization is done by subtracting the minimum value of \code{x} and dividing
the result by the specified quantile of \code{x}. Values greater than 1 are capped
at 1.
}
\examples{
x <- rnorm(100)
normed_x <- norm0q(x)

range(normed_x) # This should show values between 0 and 1

}
