% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\docType{class}
\name{PBM-class}
\alias{PBM-class}
\alias{PBM}
\alias{show,PBM-method}
\title{PBM Class}
\usage{
\S4method{show}{PBM}(object)
}
\arguments{
\item{object}{An instance of \code{PBM}.}
}
\description{
This class represents a PBM (Physical Binding Model) object.
It inherits from the IQFeature class.
}
\section{Slots}{

\describe{
\item{\code{pssm}}{A matrix representing the position-specific scoring matrix.}

\item{\code{max_energy}}{A numeric value representing the maximum energy for value normalization}

\item{\code{min_energy}}{A numeric value representing the minimal energy for value normalization}

\item{\code{energy_range}}{A numeric value representing the energy range for value normalization}

\item{\code{spat}}{A data frame representing the spatial distribution (optional)}

\item{\code{spat_min}}{A numeric value representing the spatial minimum (optional).}

\item{\code{spat_max}}{A numeric value representing the spatial maximum (optional)}

\item{\code{seq_length}}{A numeric value representing the length of the sequence for spatial distribution (optional)}
}}

