% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm.compute_local}
\alias{pbm.compute_local}
\title{Compute local PBM scores / response}
\usage{
pbm.compute_local(pbm, sequences, response = FALSE, scaling_q = NULL)
}
\arguments{
\item{pbm}{A PBM object}

\item{sequences}{A vector of DNA sequences}

\item{response}{Logical, whether to compute the response (default is FALSE)}

\item{scaling_q}{Scaling quantile for normalization (default is NULL)}
}
\value{
A matrix of computed local energies
}
\description{
This function computes local energies for a given PBM and set of sequences.
}
