% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm.gextract}
\alias{pbm.gextract}
\title{Extract PBM scores for genomic intervals}
\usage{
pbm.gextract(
  pbm,
  intervals,
  response = FALSE,
  func = "logSumExp",
  normalize_energies = TRUE
)
}
\arguments{
\item{pbm}{A PBM object}

\item{intervals}{Genomic intervals to extract}

\item{response}{Logical, whether to compute the response (default is FALSE)}

\item{func}{the function to use to combine the PWMs for each sequence. Either 'logSumExp' or 'max'. The default is 'logSumExp'.}

\item{normalize_energies}{A logical flag indicating whether to normalize the energies to a range of 0-10. Default is TRUE. Note that response computation requires normalized energies.}
}
\value{
The intervals, with an additional column containing the PBM scores / response
}
\description{
This function extracts PBM scores of a given PBM for genomic intervals and optionally computes the response.
}
