% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm.normalize_energies}
\alias{pbm.normalize_energies}
\title{Normalize energies for a PBM}
\usage{
pbm.normalize_energies(
  pbm,
  energies,
  max_energy = pbm@max_energy,
  min_energy = pbm@min_energy,
  energy_range = pbm@energy_range,
  norm_energy_max = 10
)
}
\arguments{
\item{pbm}{A PBM object}

\item{energies}{A vector of energy values to normalize}

\item{max_energy}{The maximum energy value (default is pbm@max_energy)}

\item{min_energy}{The minimum energy value (default is pbm@min_energy)}

\item{energy_range}{The range of energy values (default is pbm@energy_range)}

\item{norm_energy_max}{The maximum value for the normalized energy (default is 10)}
}
\value{
A vector of normalized energy values
}
\description{
This function normalizes the energies for a given PBM.
}
