% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm.normalize_energies_local}
\alias{pbm.normalize_energies_local}
\title{Normalize energies locally for a PBM}
\usage{
pbm.normalize_energies_local(pbm, energies)
}
\arguments{
\item{pbm}{A PBM object}

\item{energies}{A matrix of local energy values to normalize}
}
\value{
A matrix of locally normalized energy values
}
\description{
This function performs local normalization of energies for a given PBM.
}
