% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm.trim_pssm}
\alias{pbm.trim_pssm}
\title{Trim a PBM}
\usage{
pbm.trim_pssm(pbm, bits_thresh = 0.1)
}
\arguments{
\item{pbm}{A PBM object}

\item{bits_thresh}{The threshold for trimming positions (default is 0.1)}
}
\value{
A trimmed PBM object
}
\description{
This function trims a PBM object by removing positions with low information content.
}
