% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm_list.compute_local}
\alias{pbm_list.compute_local}
\title{Extract local PBM scores / response for a list of PBMs given sequences}
\usage{
pbm_list.compute_local(pbm_list, sequences, response = FALSE, scaling_q = NULL)
}
\arguments{
\item{pbm_list}{A list of PBM objects}

\item{sequences}{A vector of DNA sequences}

\item{response}{Logical, whether to compute the response (default is FALSE)}

\item{scaling_q}{Scaling quantile for normalization (default is NULL)}
}
\value{
A list of matrices containing extracted local energies for each PBM
}
\description{
This function extracts local PBM scores for sequences and optionally computes the response for a list of PBMs.
}
