% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm_list.gextract}
\alias{pbm_list.gextract}
\title{Extract PBM scores / response for a list of PBMs}
\usage{
pbm_list.gextract(pbm_list, intervals, response = FALSE, func = "logSumExp")
}
\arguments{
\item{pbm_list}{A list of PBM objects}

\item{intervals}{Genomic intervals to extract}

\item{response}{Logical, whether to compute the response (default is FALSE)}

\item{func}{the function to use to combine the PWMs for each sequence. Either 'logSumExp' or 'max'. The default is 'logSumExp'.}
}
\value{
The intervals, with an additional column containing the PBM scores / response for each PBM
}
\description{
This function extracts  PBM scores for genomic intervals and optionally computes the response for a list of PBMs.
}
