% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm_list.gextract_local}
\alias{pbm_list.gextract_local}
\title{Extract local PBM scores / response for a list of PBMs}
\usage{
pbm_list.gextract_local(
  pbm_list,
  intervals,
  response = FALSE,
  scaling_q = NULL
)
}
\arguments{
\item{pbm_list}{A list of PBM objects}

\item{intervals}{Genomic intervals to extract}

\item{response}{Logical, whether to compute the response (default is FALSE)}

\item{scaling_q}{Scaling quantile for normalization (default is NULL)}
}
\value{
A list of matrices containing extracted local energies for each PBM
}
\description{
This function extracts local PBM scores for genomic intervals and optionally computes the response for a list of PBMs.
}
