% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm_list.multi_traj.compute_energy}
\alias{pbm_list.multi_traj.compute_energy}
\title{Compute energy for a list of pbm lists (multiple trajectories)}
\usage{
pbm_list.multi_traj.compute_energy(multi_traj, sequences, func = "logSumExp")
}
\arguments{
\item{multi_traj}{A list of PBM lists}

\item{sequences}{A set of sequences on which to compute the energies.}

\item{func}{the function to use to combine the PWMs for each sequence. Either 'logSumExp' or 'max'. The default is 'logSumExp'.}
}
\value{
A matrix containing the computed energies for each PBM
}
\description{
Compute energy for a list of pbm lists (multiple trajectories)
}
