% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm_list.multi_traj.gextract_energy}
\alias{pbm_list.multi_traj.gextract_energy}
\title{Extract energy for a list of pbm lists (multiple trajectories)}
\usage{
pbm_list.multi_traj.gextract_energy(multi_traj, intervals, func = "logSumExp")
}
\arguments{
\item{multi_traj}{A list of PBM lists}

\item{intervals}{Genomic intervals to extract}

\item{func}{the function to use to combine the PWMs for each sequence. Either 'logSumExp' or 'max'. The default is 'logSumExp'.}
}
\value{
A matrix containing the computed energies for each PBM
}
\description{
Extract energy for a list of pbm lists (multiple trajectories)
}
