% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm_list.normalize}
\alias{pbm_list.normalize}
\title{Normalize energies for a list of PBMs}
\usage{
pbm_list.normalize(pbm_list, energies, norm_energy_max = 10)
}
\arguments{
\item{pbm_list}{A list of PBM objects}

\item{energies}{A matrix of energy values to normalize, with column names matching PBM names}

\item{norm_energy_max}{The maximum value for normalized energies (default is 10)}
}
\value{
A matrix of normalized energy values
}
\description{
This function normalizes energy values for a list of PBMs based on their
normalization parameters, using the pbm.normalize_energies function.
}
