% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm_list.trim_pssm}
\alias{pbm_list.trim_pssm}
\title{Trim a list of PBMs}
\usage{
pbm_list.trim_pssm(pbm_list, bits_thresh = 0.1)
}
\arguments{
\item{pbm_list}{A list of PBM objects}

\item{bits_thresh}{The threshold for trimming positions (default is 0.1)}
}
\value{
A list of trimmed PBM objects
}
\description{
Trim a list of PBMs
}
