% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{pbm_list_to_mdb}
\alias{pbm_list_to_mdb}
\title{Convert a list of PBM objects to a motif database}
\usage{
pbm_list_to_mdb(pbm_list)
}
\arguments{
\item{pbm_list}{A list of PBM objects}
}
\value{
A motif database object compatible with prego::extract_pwm
}
\description{
This function converts a list of PBM objects to a motif database format
compatible with prego::extract_pwm.
}
