% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq-plot.R
\name{plot_iq_locus}
\alias{plot_iq_locus}
\title{Plot IQ Locus}
\usage{
plot_iq_locus(
  interval,
  pbm_list,
  atac_tracks,
  width = 500,
  ext_width = 50000,
  T_emax = 8,
  bits_threshold = NULL,
  order_motifs = TRUE,
  atac_names = atac_tracks,
  atac_colors = NULL,
  atac_sizes = NULL,
  line_thresh = 0.9,
  score = NULL,
  norm_q = 0.995,
  tracks_q = NULL,
  iterator = 20,
  norm_intervals = gintervals.all(),
  atac_smooth = 2,
  ext_atac_smooth = atac_smooth * 5,
  tn5bias_track = "tn5bias",
  normalize_tn5bias = TRUE,
  tss_intervals = "intervs.global.tss",
  exon_intervals = "intervs.global.exon",
  scale_cex = 500,
  filename = NULL,
  dev = grDevices::pdf,
  plot_width = 15,
  plot_height = 8
)
}
\arguments{
\item{interval}{A genomic interval to plot.}

\item{pbm_list}{A list of PBM objects.}

\item{atac_tracks}{A vector of ATAC-seq track names to plot.}

\item{width}{Width of the plot in bp.}

\item{ext_width}{Width of the top plot in bp.}

\item{T_emax}{Numeric, threshold for maximum energy.}

\item{bits_threshold}{Numeric, threshold for trimming PSSMs.}

\item{order_motifs}{Logical, whether to order motifs by maximum response.}

\item{atac_names}{Character vector, names for ATAC-seq tracks.}

\item{atac_colors}{Named vector, colors for ATAC-seq tracks.}

\item{atac_sizes}{Named vector, line sizes for ATAC-seq tracks.}

\item{line_thresh}{Numeric, threshold for drawing vertical lines.}

\item{score}{Numeric, optional score to display in the plot title.}

\item{norm_q}{Numeric, quantile for normalization.}

\item{tracks_q}{Data frame, pre-computed quantiles for tracks. Should have a 'type' column and a 'q' column.}

\item{iterator}{Numeric, iterator for gextract.}

\item{norm_intervals}{Genomic intervals for normalization.}

\item{atac_smooth}{Numeric, smoothing factor for ATAC-seq tracks.}

\item{ext_atac_smooth}{Numeric, smoothing factor for the top plot.}

\item{tn5bias_track}{Character, name of the Tn5 bias track.}

\item{normalize_tn5bias}{Logical, whether to normalize the TN5 bias track.}

\item{scale_cex}{Numeric, scaling factor for letter sizes.}

\item{filename}{Character, output filename (if NULL, plot is not saved).}

\item{dev}{Function, device to use for plotting.}

\item{plot_width}{Numeric, width of the output plot.}

\item{plot_height}{Numeric, height of the output plot.}
}
\value{
A ggplot object containing the IQ locus plot.
}
\description{
This function generates a plot of an IQ locus, including DNA sequence, motif responses, and ATAC-seq tracks.
}
