% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_motif_energy_vs_response_boxplot}
\alias{plot_motif_energy_vs_response_boxplot}
\title{Plot Motif Energy vs Response Boxplot}
\usage{
plot_motif_energy_vs_response_boxplot(
  traj_model,
  motif,
  xlab = paste(motif, "energy"),
  ylab = "ATAC difference",
  ylim = c(-0.5, 0.5),
  fill = "lightblue1",
  title = "",
  subtitle = NULL,
  outliers = TRUE,
  y_data = traj_model@diff_score
)
}
\arguments{
\item{traj_model}{The trajectory model object.}

\item{motif}{The motif to plot the energy levels for.}

\item{xlab}{The label for the x-axis (default is the motif name followed by "energy").}

\item{ylab}{The label for the y-axis (default is "ATAC difference").}

\item{ylim}{The limits for the y-axis (default is -0.5 to 0.5).}

\item{fill}{The fill color for the boxplot (default is "lightblue1").}

\item{title}{The title for the plot.}

\item{subtitle}{The subtitle for the plot (default is a kologorov-smirnov test between the lowest and highest energy levels). The color of the subtitle will be set to "darkred" if the p-value is less than 0.01.}

\item{outliers}{Whether to plot outliers (default is TRUE).}

\item{y_data}{The y data to use for the plot (default is the ATAC difference).}
}
\value{
A ggplot object representing the boxplot.
}
\description{
This function plots a boxplot of the ATAC difference (response) against the energy levels of a given motif.
If \code{subtitle=NULL}, the subtitle will be set to a kologorov-smirnov test between the lowest (0-3) and highest (9-10) energy levels.
}
