% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot_motif_spatial_freq}
\alias{plot_motif_spatial_freq}
\title{Plot motif spatial frequency}
\usage{
plot_motif_spatial_freq(spatial_freqs, motif, smooth = 10, plot_atac = FALSE)
}
\arguments{
\item{spatial_freqs}{A data frame containing the spatial frequency of motifs and/or ATAC-seq peaks. Output of \code{compute_traj_model_spatial_freq}.}

\item{motif}{A string specifying the motif to plot.}

\item{smooth}{An integer specifying the window size for smoothing the frequency values.}

\item{plot_atac}{A logical indicating whether to plot the ATAC-seq frequency instead of the motif frequency.}
}
\value{
A ggplot object showing the spatial frequency of the given motif.
}
\description{
This function plots the spatial frequency of a given motif across genomic positions.
}
