% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_motifs_report}
\alias{plot_motifs_report}
\title{Plot a report of trajectory motifs}
\usage{
plot_motifs_report(
  traj_model,
  motif_num = NULL,
  free_coef_axis = TRUE,
  filename = NULL,
  width = NULL,
  height = NULL,
  dev = grDevices::pdf,
  ...
)
}
\arguments{
\item{traj_model}{Trajectory model object. Please run \code{regress_trajectory_motifs} first.}

\item{motif_num}{Number of motifs to plot. If NULL, all motifs will be plotted.}

\item{free_coef_axis}{Whether to use a free axis for the coefficient plots.}

\item{filename}{Filename to save the plot to. If NULL, the plot will be returned.}

\item{width}{Width of the plot.}

\item{height}{Height of the plot.}

\item{dev}{Device to use for saving the plot. Default: \code{grDevices::pdf}.}

\item{...}{Additional arguments to pass to the device.}
}
\value{
ggplot2 object. If filename is not NULL, the plot will be saved to the file and the function will return \code{invisible(NULL)}.
}
\description{
Plot a report of trajectory motifs
}
