% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-multi-traj.R
\name{plot_multi_traj_model_report}
\alias{plot_multi_traj_model_report}
\title{Plot multi-trajectory model report}
\usage{
plot_multi_traj_model_report(
  multi_traj,
  filename = NULL,
  width = NULL,
  height = NULL,
  dev = grDevices::pdf,
  title = NULL,
  use_full = FALSE,
  names_map = NULL,
  type = "pr",
  ...
)
}
\arguments{
\item{multi_traj}{The multi-trajectory model object.}

\item{filename}{Optional. The name of the output file. If not provided, the plot will be displayed on the screen.}

\item{width}{Optional. The width of the output plot in inches.}

\item{height}{Optional. The height of the output plot in inches.}

\item{dev}{Optional. The graphics device to use for output. Default is 'pdf'.}

\item{title}{Optional. The title of the plot.}

\item{use_full}{Optional. Whether to use the models with all the motifs or not. Default is FALSE.}

\item{names_map}{a named vector to map the names of the motifs to new names.}

\item{type}{Optional. The type of the plot, can be "pr", "boxp" or "spat". Default is "pr".}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
A ggplot2 object (invsibly if a filename is provided).
}
\description{
This function plots a report for a multi-trajectory model. The report includes the motifs, spatial models, and partial responses for each trajectory model when \code{type} is "pr".
If \code{type} is "boxp", the function will plot the boxplots of the energy vs response for each motif. If \code{type} is "spat", the function will plot the spatial distribution of the motifs.
}
