% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_partial_response}
\alias{plot_partial_response}
\title{Plot Partial Response}
\usage{
plot_partial_response(
  traj_model,
  motif,
  ylim = NULL,
  xlab = "Energy",
  ylab = "Partial response",
  pointsize = 3
)
}
\arguments{
\item{traj_model}{The trajectory model object.}

\item{motif}{The motif to plot the partial response for.}

\item{ylim}{The y-axis limits for the plot (optional).}

\item{xlab}{The label for the x-axis (optional, default is "Energy").}

\item{ylab}{The label for the y-axis (optional, default is "Partial response").}

\item{pointsize}{The size of the points in the scatter plot (optional, default is 3).}
}
\value{
A ggplot object representing the partial response plot.
}
\description{
This function plots the partial response for a given trajectory model and motif.
}
