% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_prediction_boxplot}
\alias{plot_prediction_boxplot}
\title{Plot a boxplot of observed vs predicted ATAC differences}
\usage{
plot_prediction_boxplot(traj_model, n_groups = 5)
}
\arguments{
\item{traj_model}{A trajectory model object}

\item{n_groups}{The number of groups to divide the observed ATAC difference into}
}
\value{
A ggplot object representing the boxplot
}
\description{
This function takes a trajectory model and plots a boxplot of the observed vs predicted ATAC differences.
The plot is grouped by the observed ATAC difference, which is divided into n_groups quantiles.
}
