% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_traj_model_clusters_report}
\alias{plot_traj_model_clusters_report}
\title{Plot trajectory model clusters report}
\usage{
plot_traj_model_clusters_report(traj_model, dir, k = 10, spatial_freqs = NULL)
}
\arguments{
\item{traj_model}{A trajectory model object}

\item{dir}{A directory to save the report files}

\item{k}{Number of clusters to split the heatmap into}

\item{spatial_freqs}{A vector of spatial frequencies to plot. Use \code{compute_traj_model_spatial_freq} to compute.}
}
\value{
None. A file called \code{heatmap.png} with the heatmap and a file called \code{motifs_report.pdf} with the motifs report will be saved to the directory.
}
\description{
Plot trajectory model clusters report
}
