% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_variable_vs_response}
\alias{plot_variable_vs_response}
\title{Plot scatter of a variable vs ATAC difference}
\usage{
plot_variable_vs_response(traj_model, variable, point_size = 0.5)
}
\arguments{
\item{traj_model}{Trajectory model object. Please run \code{\link{regress_trajectory_motifs}} first.}

\item{variable}{Variable to plot.}

\item{point_size}{Size of the points.}
}
\value{
ggplot2 object
}
\description{
Plot scatter of a variable vs ATAC difference
}
