% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-regression.R
\name{regress_trajectory_motifs}
\alias{regress_trajectory_motifs}
\title{Perform motif regression on ATAC trajectories.}
\usage{
regress_trajectory_motifs(
  peak_intervals,
  atac_scores = NULL,
  atac_diff = NULL,
  normalize_bins = TRUE,
  norm_intervals = NULL,
  max_motif_num = 30,
  n_clust_factor = 1,
  motif_energies = NULL,
  norm_motif_energies = NULL,
  pssm_db = iceqream::motif_db,
  additional_features = NULL,
  min_tss_distance = 5000,
  bin_start = 1,
  bin_end = NULL,
  min_initial_energy_cor = 0.05,
  normalize_energies = TRUE,
  energy_norm_quantile = 1,
  norm_energy_max = 10,
  n_prego_motifs = 0,
  traj_prego = NULL,
  min_diff = 0.1,
  distill_on_diff = FALSE,
  prego_sample_for_kmers = TRUE,
  prego_sample_fraction = 0.1,
  seed = 60427,
  feature_selection_beta = 0.003,
  lambda = 0.00001,
  alpha = 1,
  filter_using_r2 = FALSE,
  r2_threshold = 0.0005,
  parallel = TRUE,
  peaks_size = 500,
  spat_num_bins = NULL,
  spat_bin_size = 2,
  kmer_sequence_length = 300,
  include_interactions = FALSE,
  interaction_threshold = 0.001,
  max_motif_interaction_n = NULL,
  max_add_interaction_n = NULL
)
}
\arguments{
\item{peak_intervals}{A data frame, indicating the genomic positions ('chrom', 'start', 'end') of each peak.}

\item{atac_scores}{Optional. A numeric matrix, representing mean ATAC score per bin per peak. Rows: peaks, columns: bins. By default iceqream would regress the last column minus the first column. If you want to regress something else, please either change bin_start or bin_end, or provide \code{atac_diff} instead. If \code{normalize_bins} is TRUE, the scores will be normalized to be between 0 and 1.}

\item{atac_diff}{Optional. A numeric vector representing the differential accessibility between the start and end of the trajectory. Either this or \code{atac_scores} must be provided.}

\item{normalize_bins}{whether to normalize the ATAC scores to be between 0 and 1. Default: TRUE}

\item{norm_intervals}{A data frame, indicating the genomic positions ('chrom', 'start', 'end') of peaks used for energy normalization. If NULL, the function will use \code{peak_intervals} for normalization.}

\item{max_motif_num}{maximum number of motifs to consider. Default: 50}

\item{n_clust_factor}{factor to divide the number of to keep after clustering. e.g. if n_clust_factor > 1 the number of motifs to keep will be reduced by a factor of n_clust_factor. Default: 1}

\item{motif_energies}{A numeric matrix, representing the energy of each motif in each peak. If NULL, the function will use \code{pssm_db} to calculate the motif energies. Note that this might take a while.}

\item{norm_motif_energies}{A numeric matrix, representing the normalized energy of each motif in each interval of \code{norm_intervals}. If NULL, the function will use \code{pssm_db} to calculate the motif energies. Note that this might take a while.}

\item{pssm_db}{a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name. All the motifs in \code{motif_energies} (column names) should be present in the 'motif' column. Default: all motifs.}

\item{additional_features}{A data frame, representing additional genomic features (e.g. CpG content, distance to TSS, etc.) for each peak. Note that NA values would be replaced with 0.}

\item{min_tss_distance}{distance from Transcription Start Site (TSS) to classify a peak as an enhancer. Default: 5000. If NULL, no filtering will be performed - use this option if your peaks are already filtered. \cr
Note that in order to filter peaks that are too close to TSS, the current \code{misha} genome must have an intervals set called \code{intervs.global.tss}.}

\item{bin_start}{the start of the trajectory. Default: 1}

\item{bin_end}{the end of the trajectory. Default: the last bin (only used when atac_scores is provided)}

\item{min_initial_energy_cor}{minimal correlation between the motif normalized energy and the ATAC difference.}

\item{normalize_energies}{whether to normalize the motif energies. Set this to FALSE if the motif energies are already normalized.}

\item{energy_norm_quantile}{quantile of the energy used for normalization. Default: 1}

\item{norm_energy_max}{maximum value of the normalized energy. Default: 10}

\item{n_prego_motifs}{number of prego motifs (de-novo motifs) to consider.}

\item{traj_prego}{output of \code{learn_traj_prego}. If provided, no additional prego models would be inferred.}

\item{min_diff}{minimal ATAC difference for a peak to participate in the initial prego motif inference and in the distillation step (if \code{distill_on_diff} is TRUE).}

\item{distill_on_diff}{whether to distill motifs based on differential accessibility. If FALSE, all peaks will be used for distillation, if TRUE - only peaks with differential accessibility >= min_diff will be used.}

\item{prego_sample_for_kmers}{whether to use a sample of the peaks for kmer screening. Default: TRUE}

\item{prego_sample_fraction}{Fraction of peaks to sample for prego motif inference. A smaller number would be faster but might lead to over-fitting. Default: 0.1}

\item{seed}{random seed for reproducibility.}

\item{feature_selection_beta}{beta parameter used for feature selection.}

\item{lambda}{A user supplied \code{lambda} sequence. Typical usage is to
have the program compute its own \code{lambda} sequence based on
\code{nlambda} and \code{lambda.min.ratio}. Supplying a value of
\code{lambda} overrides this. WARNING: use with care. Avoid supplying a
single value for \code{lambda} (for predictions after CV use
\code{predict()} instead).  Supply instead a decreasing sequence of
\code{lambda} values. \code{glmnet} relies on its warms starts for speed,
and its often faster to fit a whole path than compute a single fit.}

\item{alpha}{The elasticnet mixing parameter, with \eqn{0\le\alpha\le 1}.
The penalty is defined as
\deqn{(1-\alpha)/2||\beta||_2^2+\alpha||\beta||_1.} \code{alpha=1} is the
lasso penalty, and \code{alpha=0} the ridge penalty.}

\item{filter_using_r2}{whether to filter features using R^2.}

\item{r2_threshold}{minimal R^2 for a feature to be included in the model.}

\item{parallel}{whether to use parallel processing on glmnet.}

\item{peaks_size}{size of the peaks to extract sequences from. Default: 300bp}

\item{spat_num_bins}{number of spatial bins to use.}

\item{spat_bin_size}{size of each spatial bin.}

\item{kmer_sequence_length}{length of the kmer sequence to use for kmer screening. By default the full sequence is used.}

\item{include_interactions}{whether to include interactions between motifs / additional fetures as model features. IQ will create interactions between significant additional features and all motifs, and between significant motifs. Default: FALSE}

\item{interaction_threshold}{threshold for the selecting features to create interactions. IQ learns a linear model on the features and selects the features with coefficients above this threshold. Default: 0.001}

\item{max_motif_interaction_n}{maximum number of motifs to consider for interactions. If NULL, all motifs above the interaction_threshold will be considered. Default: NULL}

\item{max_add_interaction_n}{maximum number of additional features to consider for interactions. If NULL, all additional features above the interaction_threshold will be considered. Default: NULL}
}
\value{
An instance of \code{TrajectoryModel} containing model information and results:
\itemize{
\item model: The final General Linear Model (GLM) object.
\item motif_models: Named List, PSSM and spatial models for each motif cluster.
\item normalized_energies: Numeric vector, normalized energies of each motif in each peak.
\item additional_features: data frame of the additional features.
\item diff_score: Numeric, normalized score of differential accessibility between 'bin_start' and 'bin_end'.
\item predicted_diff_score: Numeric, predicted differential accessibility score between 'bin_start' and 'bin_end'.
\item initial_prego_models: List, inferred prego models at the initial step of the algorithm.
\item peak_intervals: data frame, indicating the genomic positions ('chrom', 'start', 'end') of each peak used for training.
}
Print the model to see more details.
}
\description{
This function performs motif regression on ATAC trajectories. It can work with either ATAC scores on trajectory bins or directly with differential accessibility.

The basic inputs for regression are the genomic positions of the peaks, two vectors of ATAC scores (as an \code{atac_scores} matrix) or differential accessibility (\code{atac_diff}), and database energies computed for all the genomics positions.
Optionally, additional features such as epigenomic features can be provided.
}
