% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj-model-utils.R
\name{relearn_traj_model}
\alias{relearn_traj_model}
\title{Relearn Trajectory Model}
\usage{
relearn_traj_model(
  traj_model,
  new_energies = FALSE,
  new_logist = FALSE,
  lambda = NULL,
  verbose = FALSE
)
}
\arguments{
\item{traj_model}{The trajectory model object.}

\item{new_energies}{If TRUE - recreate the energies. Default is FALSE.}

\item{new_logist}{If TRUE - recreate the logistic features. Default is FALSE. If \code{new_energies} is TRUE, this is automatically set to TRUE.}

\item{lambda}{The lambda value to use for relearning. If NULL, the lambda value from the trajectory model is used.}

\item{verbose}{Logical indicating whether to display additional information.}
}
\value{
The updated trajectory model object.
}
\description{
This function relearns a trajectory model using the glmnet package.
}
