% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-model.R
\name{remove_motif_models_from_traj}
\alias{remove_motif_models_from_traj}
\title{Remove motif models from trajectory model}
\usage{
remove_motif_models_from_traj(traj_model, motif_models, verbose = TRUE)
}
\arguments{
\item{traj_model}{The trajectory model object to remove motif models from.}

\item{motif_models}{A character vector specifying the names of the motif models to remove.}

\item{verbose}{A logical value indicating whether to display information about the R^2 after removing the motif models. Default is TRUE.}
}
\value{
The updated trajectory model object after removing the motif models.
}
\description{
This function removes specified motif models from a given trajectory model.
It updates the model, motif models, predicted difference score, model features,
coefficients, normalized energies, and features R^2 of the trajectory model.
}
