% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj-model-utils.R
\name{rename_motif_models}
\alias{rename_motif_models}
\title{Rename motif models in a trajectory model}
\usage{
rename_motif_models(traj_model, names_map)
}
\arguments{
\item{traj_model}{The trajectory model object to modify.}

\item{names_map}{A named character vector specifying the mapping of current motif model names to new names.
Can be generated using the \code{match_traj_model_motif_names} function.}
}
\value{
The modified trajectory model object with renamed motif models.
}
\description{
This function renames motif models in a trajectory model based on a provided names_map.
The names_map should be a named character vector where the names are the current names of the motif models
and the values are the new names.
}
