% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQFeature.R
\name{show,IQFeature-method}
\alias{show,IQFeature-method}
\title{Show method for IQFeature}
\usage{
\S4method{show}{IQFeature}(object)
}
\arguments{
\item{object}{An IQFeature object}
}
\description{
This method defines how an IQFeature object should be displayed.
}
