% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQFeatureGroup.R
\name{show,IQFeatureGroup-method}
\alias{show,IQFeatureGroup-method}
\title{Show method for IQFeatureGroup}
\usage{
\S4method{show}{IQFeatureGroup}(object)
}
\arguments{
\item{object}{An IQFeatureGroup object}
}
\description{
This method defines how an IQFeatureGroup object should be displayed.
}
