% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQModel.R
\name{show,IQmodel-method}
\alias{show,IQmodel-method}
\title{Show method for IQmodel}
\usage{
\S4method{show}{IQmodel}(object)
}
\arguments{
\item{object}{An IQmodel object}
}
\description{
This method defines how an IQmodel object should be displayed.
}
