% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQSeqFeature.R
\name{show,IQSeqFeature-method}
\alias{show,IQSeqFeature-method}
\title{Show method for IQSeqFeature}
\usage{
\S4method{show}{IQSeqFeature}(object)
}
\arguments{
\item{object}{An IQSeqFeature object}
}
\description{
This method defines how an IQSeqFeature object should be displayed.
}
