% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj-model-utils.R
\name{split_traj_model_to_train_test}
\alias{split_traj_model_to_train_test}
\title{Split a trajectory model into train and test sets}
\usage{
split_traj_model_to_train_test(traj_model)
}
\arguments{
\item{traj_model}{The trajectory model to split}
}
\value{
A list with the train and test sets
}
\description{
This function takes a trajectory model and splits it into train and test sets.
}
