% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-utils.R
\name{strip_traj_model}
\alias{strip_traj_model}
\title{Strip a trajectory model to essential components}
\usage{
strip_traj_model(traj_model)
}
\arguments{
\item{traj_model}{A TrajectoryModel object}
}
\value{
A stripped TrajectoryModel object
}
\description{
Strip a trajectory model to essential components
}
