% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQInteraction.R
\name{traj_model_interactions_to_iq_feature_list}
\alias{traj_model_interactions_to_iq_feature_list}
\title{Convert trajectory model interactions to IQ feature list}
\usage{
traj_model_interactions_to_iq_feature_list(traj_model)
}
\arguments{
\item{traj_model}{The trajectory model object to convert.}
}
\value{
A list of IQ interaction features, where each feature contains the interaction name and its corresponding coefficients.
}
\description{
This function takes a trajectory model with interactions and converts them to a list of IQ interaction features.
}
