% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQFeature.R
\name{traj_model_to_iq_feature_list}
\alias{traj_model_to_iq_feature_list}
\title{Convert trajectory model to IQ feature list}
\usage{
traj_model_to_iq_feature_list(traj_model)
}
\arguments{
\item{traj_model}{The trajectory model object to convert.}
}
\value{
A list of IQ features, where each feature contains the feature name and its corresponding coefficients.
}
\description{
This function takes a trajectory model object and converts its additional features into a list of IQ features.
}
