% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBM.R
\name{traj_model_to_pbm_list}
\alias{traj_model_to_pbm_list}
\title{Convert a trajectory model to a list of PBM objects}
\usage{
traj_model_to_pbm_list(
  traj_model,
  func = "logSumExp",
  normalization_intervals = traj_model@normalization_intervals,
  bits_threshold = NULL
)
}
\arguments{
\item{traj_model}{A trajectory model object}

\item{func}{The function to use for computing the energies "logSumExp" or "max". Default is "logSumExp"}

\item{normalization_intervals}{The normalization intervals to use for computing the energies. Default is the normalization intervals of the trajectory model.}

\item{bits_threshold}{The threshold for trimming the PSSM (default is NULL, no trimming)}
}
\value{
A list of PBM objects
}
\description{
This function takes a trajectory model and converts it into a list of PBM objects.
}
