% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-response.R
\name{traj_model_variable_response}
\alias{traj_model_variable_response}
\title{Compute Partial Response for Trajectory Model}
\usage{
traj_model_variable_response(
  traj_model,
  vars = NULL,
  lambda = 0.00001,
  reverse = FALSE
)
}
\arguments{
\item{traj_model}{A trajectory model object.}

\item{vars}{A vector of variables to compute the partial response for. Default is NULL (all variables).}

\item{lambda}{A regularization parameter. Default is 1e-5.}

\item{reverse}{If TRUE, the partial response is computed for all variables except the ones in \code{vars}. Default is FALSE.}
}
\value{
A data frame with the computed partial responses, where the column names are the variables.
}
\description{
This function computes the partial response for a given trajectory model.
}
