% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-regression.R
\docType{class}
\name{TrajectoryModel-class}
\alias{TrajectoryModel-class}
\alias{TrajectoryModel}
\title{TrajectoryModel Class}
\description{
An S4 class to represent a trajectory model.
}
\section{Slots}{

\describe{
\item{\code{diff_score}}{numeric
A numerical value representing the difference score.}

\item{\code{predicted_diff_score}}{numeric
A numerical value representing the predicted difference score.}

\item{\code{model}}{ANY
The model object or any data type that represents the trajectory model.}

\item{\code{normalized_energies}}{matrix
A matrix containing normalized energies.}

\item{\code{motif_models}}{list
A list of models representing different motifs.}

\item{\code{initial_prego_models}}{list
A list of initial pre-go models.}

\item{\code{train_peak_intervals}}{data.frame
A data frame containing the peak intervals used for training.}

\item{\code{params}}{list
A list of parameters used for training.}

\item{\code{additional_features}}{character
A character vector of the names of additional features.}
}}

