% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryModelMulti.R
\docType{class}
\name{TrajectoryModelMulti-class}
\alias{TrajectoryModelMulti-class}
\alias{TrajectoryModelMulti}
\alias{show,TrajectoryModelMulti-method}
\title{TrajectoryModelMulti Class}
\usage{
\S4method{show}{TrajectoryModelMulti}(object)
}
\arguments{
\item{object}{An instance of \code{TrajectoryModelMulti}.}
}
\description{
This class represents a multi-trajectory model, which is a collection of individual trajectory models.
}
\section{Slots}{

\describe{
\item{\code{models}}{A list of individual trajectory models.}

\item{\code{models}}{A list of individual trajectory models with full features.}

\item{\code{motif_models}}{A list of motif models corresponding to the individual trajectory models.}

\item{\code{cluster_map}}{a data frame that maps original motif to its distilled cluster.}

\item{\code{stats}}{A data frame with r2 statistics for each trajectory model after distillation.}

\item{\code{params}}{additional parameters for the multi-trajectory distillation.}
}}

