% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{compute_traj_model_spatial_freq}
\alias{compute_traj_model_spatial_freq}
\title{Compute spatial frequency of motifs in trajectory model}
\usage{
compute_traj_model_spatial_freq(
  traj_model,
  size,
  pwm_threshold = 7,
  top_q = 0.1,
  bottom_q = 0.1,
  atac_track = NULL,
  parallel = TRUE,
  bidirect_size = NULL,
  k4me3_track = NULL,
  k27me3_track = NULL,
  k27ac_track = NULL,
  orient_to_intervals = NULL,
  align_to_max = TRUE
)
}
\arguments{
\item{traj_model}{A trajectory model object}

\item{size}{The size of the region to compute spatial frequency for}

\item{pwm_threshold}{The threshold for the PWM score}

\item{top_q}{The proportion of top peaks to select}

\item{bottom_q}{The proportion of bottom peaks to select}

\item{bidirect_size}{Size of the intervals to use for deciding the directionality of the sequence}

\item{k4me3_track}{name of k4me3 track}

\item{k27me3_track}{name of k27me3 track}

\item{k27ac_track}{name of k27ac track}

\item{orient_to_intervals}{A data frame containing the intervals to orient the sequences to}

\item{align_to_max}{A logical indicating whether to align the sequences to the maximum in each sequence at the epigenetic tracks and atac signla}
}
\value{
A data frame with the spatial frequency of each motif
}
\description{
This function computes the spatial frequency of motifs in a trajectory model, using the top and bottom 10\% of peaks based on diff_score.
}
