% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{create_default_additional_features}
\alias{create_default_additional_features}
\title{Create Default Additional Features}
\usage{
create_default_additional_features(
  peaks,
  tracks,
  window_size = 20000,
  normalize = TRUE,
  norm_quant = 0.05,
  spatial_ratio_ext = 1000
)
}
\arguments{
\item{peaks}{Intervals set of peaks}

\item{tracks}{Tracks containing ATAC signal}

\item{window_size}{An integer specifying the size of the window for computing regional ATAC signal. Default is 20,000 bp.}

\item{normalize}{A logical value indicating whether to normalize the features to the range 0-10.}

\item{norm_quant}{The quantile to use for normalization. Values below this quantile and above 1-quantile are truncated, and the rest are linearly scaled to 0-10.}

\item{spatial_ratio_ext}{Extension size (in bp) for computing the spatial ratio feature. Default is 1000.}
}
\value{
A data frame with additional features: sequence features and regional ATAC signal. The features are normalized to the range 0-10.
}
\description{
This function generates additional features for a given set of peaks. The features include sequence features and regional ATAC signal using a punctured window.
}
