% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQSeqFeature.R
\name{create_dinuc_feature_group}
\alias{create_dinuc_feature_group}
\title{Create Dinucleotide Feature Group}
\usage{
create_dinuc_feature_group(
  traj_model,
  quantile = 0.95,
  size = NULL,
  dinucleotides = c("AA", "AC", "AG", "AT", "CA", "CC", "CG", "CT", "GA", "GC", "GG",
    "GT", "TA", "TC", "TG", "TT"),
  include_gc = TRUE
)
}
\arguments{
\item{traj_model}{A trajectory model object.}

\item{quantile}{The quantile to use for normalization (default is 0.95).}

\item{size}{The size of the sequences to use for feature computation.}

\item{dinucleotides}{A vector of dinucleotides to create features for (default is all possible dinucleotides).}

\item{include_gc}{Logical indicating whether to include GC content feature (default is TRUE).}
}
\value{
An IQFeatureGroup object containing all dinucleotide features and GC content if included.
}
\description{
This function creates an IQFeatureGroup object for all dinucleotide features and GC content.
}
