% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill-motifs.R
\name{distill_traj_model}
\alias{distill_traj_model}
\title{Distill trajectory model}
\usage{
distill_traj_model(traj_model, max_motif_num, parallel = TRUE)
}
\arguments{
\item{traj_model}{A trajectory model object}

\item{max_motif_num}{The maximum number of motifs to select}

\item{parallel}{Whether to use parallel processing}
}
\value{
A distilled trajectory model object with \code{max_motif_num} models + additional features.
}
\description{
This function takes a trajectory model and reduces the number of motifs using 'prego'.
}
