% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceqream-package.R
\docType{package}
\name{iceqream-package}
\alias{iceqream}
\alias{iceqream-package}
\title{iceqream: Interpretable Computational Engine for Quantitative Regression of Enhancer ATAC Motifs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

Authors:
\itemize{
  \item Akhiad Bercovich \email{akhiad.bercovich@weizmann.ac.il}
  \item Amos Tanay \email{amos.tanay@weizmann.ac.il}
}

}
\keyword{internal}
