% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{infer_trajectory_motifs}
\alias{infer_trajectory_motifs}
\title{Infer trajectory motifs using a pre-trained trajectory model}
\usage{
infer_trajectory_motifs(
  traj_model,
  peak_intervals,
  atac_scores = NULL,
  bin_start = 1,
  bin_end = ncol(atac_scores),
  additional_features = NULL
)
}
\arguments{
\item{traj_model}{A trajectory model object, as returned by \code{regress_trajectory_motifs}}
}
\value{
a \code{TrajectoryModel} object which contains both the original ('train') peaks and the newly inferred ('test') peaks. The field \verb{@type} indicates whether a peak is a 'train' or 'test' peak.
}
\description{
This function infers the motif energies for a set of peaks using a pre-trained trajectory model.
}
