% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill-multi-traj.R
\name{infer_trajectory_motifs_multi}
\alias{infer_trajectory_motifs_multi}
\title{Infer trajectory motifs for a multi-trajectory model}
\usage{
infer_trajectory_motifs_multi(
  traj_multi,
  peak_intervals,
  atac_scores = NULL,
  bin_start = 1,
  bin_end = ncol(atac_scores),
  additional_features = NULL
)
}
\arguments{
\item{traj_multi}{A TrajectoryModelMulti object.}

\item{peak_intervals}{A data frame, indicating the genomic positions ('chrom', 'start', 'end') of each peak, with an additional column named "const" indicating whether the peak is constitutive. Optionally, a column named "cluster" can be added with indication of the cluster of each peak.}

\item{atac_scores}{A numeric matrix, representing mean ATAC score per bin per peak. Rows: peaks, columns: bins.}

\item{bin_start}{the start of the trajectory. Default: 1}

\item{bin_end}{the end of the trajectory. Default: the last bin}

\item{additional_features}{A data frame, representing additional genomic features (e.g. CpG content, distance to TSS, etc.) for each peak. Note that NA values would be replaced with 0.}
}
\value{
A TrajectoryModelMulti object with the result of \code{infer_trajectory_motifs} for each individual trajectory.
}
\description{
This function infers trajectory motifs for each individual trajectory in a multi-trajectory model.
}
