% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQFeature.R
\name{iq_feature_list.compute}
\alias{iq_feature_list.compute}
\title{Compute features for a list of IQFeature and IQFeatureGroup objects}
\usage{
iq_feature_list.compute(feature_list, sequences = NULL, intervals = NULL)
}
\arguments{
\item{feature_list}{A list of IQFeature and/or IQFeatureGroup objects.}

\item{sequences}{A vector of DNA sequences. Optional if intervals are provided.}

\item{intervals}{A data frame of genomic intervals with columns 'chrom', 'start', and 'end'. Optional if sequences are provided.}
}
\value{
A data frame containing the computed features for all objects in the feature_list.
}
\description{
This function computes features for a given list of IQFeature and IQFeatureGroup objects,
using either provided sequences or intervals.
}
