% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQModel.R
\name{iq_model.compute_features}
\alias{iq_model.compute_features}
\title{Compute all features or responses required by an IQModel}
\usage{
iq_model.compute_features(
  model,
  sequences = NULL,
  intervals = NULL,
  new_data = NULL,
  return_responses = FALSE,
  return_separate = FALSE,
  normalize_energies = TRUE
)
}
\arguments{
\item{model}{An IQModel object}

\item{sequences}{Optional vector of DNA sequences}

\item{intervals}{Optional data frame of genomic intervals}

\item{new_data}{Optional data frame of pre-computed features}

\item{return_responses}{Logical indicating whether to return responses (TRUE) or raw features (FALSE)}

\item{return_separate}{Logical indicating whether to return separate data frames}

\item{normalize_energies}{Logical indicating whether to normalize PBM energies}
}
\value{
A data frame of computed features/responses or list of separate components
}
\description{
This function computes features or feature responses for an IQModel based on
the provided inputs.
}
