% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQFeature.R
\docType{class}
\name{IQFeature-class}
\alias{IQFeature-class}
\alias{IQFeature}
\alias{show,IQFeature-method}
\title{IQFeature class}
\usage{
\S4method{show}{IQFeature}(object)
}
\arguments{
\item{object}{An IQFeature object}
}
\description{
This class represents an IQ feature

This method defines how an IQFeature object should be displayed.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the IQ feature (character).}

\item{\code{coefs}}{The coefficients of the IQ feature (numeric).}
}}

