% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prego.R
\name{learn_traj_prego}
\alias{learn_traj_prego}
\title{Learn 'prego' models for ATAC difference of a trajectory}
\usage{
learn_traj_prego(
  peak_intervals,
  atac_diff,
  n_motifs,
  min_diff = 0.2,
  energy_norm_quantile = 1,
  min_energy = -10,
  sample_fraction = 0.1,
  sequences = NULL,
  seed = NULL,
  peaks_size = 300,
  additional_features = NULL,
  ...
)
}
\arguments{
\item{peak_intervals}{A data frame, indicating the genomic positions ('chrom', 'start', 'end') of each peak, with an additional column named "const" indicating whether the peak is constitutive. Optionally, a column named "cluster" can be added with indication of the cluster of each peak.}

\item{atac_diff}{A numeric vector, indicating the ATAC difference of each peak}

\item{n_motifs}{Number of motifs to learn}

\item{min_diff}{Minimum ATAC difference to include a peak in the training}

\item{energy_norm_quantile}{quantile of the energy used for normalization. Default: 1}

\item{min_energy}{Minimum energy value after normalization (default: -10)}

\item{sample_fraction}{Fraction of peaks to sample for training. Default: 0.1}

\item{sequences}{A character vector of sequences to learn the motifs on. If NULL, the sequences of the peaks are used.}

\item{seed}{Random seed}

\item{peaks_size}{size of the peaks to extract sequences from. Default: 300bp}

\item{additional_features}{A matrix of additional features to filter out before learning the motifs (e.g. CpG content, dinucleotide content, etc.)}

\item{...}{Additional arguments to be passed to \code{prego::regress_pwm}}
}
\description{
Learn 'prego' models for ATAC difference of a trajectory
}
