% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-utils.R
\name{norm_energy}
\alias{norm_energy}
\title{Normalize energy of a vector}
\usage{
norm_energy(x, min_energy = -10, q = 0.99)
}
\arguments{
\item{x}{A vector of energy values}

\item{min_energy}{Minimum energy value}

\item{q}{Quantile of the energies to consider as maximum}
}
\description{
Normalize energy of a vector
}
\examples{
x <- runif(n = 100, min = -11, max = 0)
norm_energy(x)

}
