% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-utils.R
\name{normalize_with_db_quantiles}
\alias{normalize_with_db_quantiles}
\title{Normalize motif energies using pre-computed quantiles}
\usage{
normalize_with_db_quantiles(
  motif_energies,
  db_quantiles,
  energy_norm_quantile,
  min_energy,
  norm_energy_max
)
}
\arguments{
\item{motif_energies}{Matrix of motif energies to normalize}

\item{db_quantiles}{Matrix of pre-computed quantiles for normalization}

\item{energy_norm_quantile}{Quantile to use for normalization}

\item{min_energy}{Minimum energy value}

\item{norm_energy_max}{Maximum normalized energy value}
}
\value{
Normalized motif energies matrix
}
\description{
Normalize motif energies using pre-computed quantiles
}
