% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQModel.R
\name{predict,IQmodel-method}
\alias{predict,IQmodel-method}
\title{Predict method for IQmodel}
\usage{
\S4method{predict}{IQmodel}(
  object,
  new_data = NULL,
  intervals = NULL,
  sequences = NULL,
  pbm_responses = NULL,
  iq_responses = NULL
)
}
\arguments{
\item{object}{An IQmodel object}

\item{new_data}{Optional data frame or matrix of new data for IQ features}

\item{intervals}{Optional genomic intervals}

\item{sequences}{Optional DNA sequences}

\item{pbm_responses}{Optional pre-computed PBM responses}

\item{iq_responses}{Optional pre-computed IQ responses}
}
\value{
A vector of normalized and rescaled predictions
}
\description{
This method makes predictions using an IQmodel object on new data.
}
