% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-regression.R
\name{regress_trajectory_motifs}
\alias{regress_trajectory_motifs}
\title{Perform motif regression on ATAC trajectories.}
\usage{
regress_trajectory_motifs(
  atac_scores,
  peak_intervals,
  max_motif_num = 50,
  motif_energies = NULL,
  pssm_db = prego::all_motif_datasets(),
  additional_features = NULL,
  min_tss_distance = 5000,
  bin_start = 1,
  bin_end = ncol(atac_scores),
  min_initial_energy_cor = 0.05,
  normalize_energies = TRUE,
  energy_norm_quantile = 1,
  n_prego_motifs = 4,
  traj_prego = NULL,
  min_diff = 0.2,
  prego_sample_fraction = 0.1,
  seed = 60427,
  feature_selection_beta = 0.003,
  lambda = 0.00001,
  alpha = 1,
  filter_using_r2 = FALSE,
  r2_threshold = 0.0005,
  parallel = TRUE,
  peaks_size = 300,
  spat_num_bins = NULL,
  spat_bin_size = NULL,
  kmer_sequence_length = NULL
)
}
\arguments{
\item{atac_scores}{A numeric matrix, representing mean ATAC score per bin per peak. Rows: peaks, columns: bins.}

\item{peak_intervals}{A data frame, indicating the genomic positions ('chrom', 'start', 'end') of each peak, with an additional column named "const" indicating whether the peak is constitutive. Optionally, a column named "cluster" can be added with indication of the cluster of each peak.}

\item{max_motif_num}{maximum number of motifs to consider. Default: 50}

\item{motif_energies}{A numeric matrix, representing the energy of each motif in each peak. If NULL, the function will use \code{pssm_db} to calculate the motif energies. Note that this might take a while.}

\item{pssm_db}{a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name. All the motifs in \code{motif_energies} (column names) should be present in the 'motif' column. Default: all motifs in the prego package.}

\item{additional_features}{A data frame, representing additional genomic features (e.g. CpG content, distance to TSS, etc.) for each peak. Note that NA values would be replaced with 0.}

\item{min_tss_distance}{distance from Transcription Start Site (TSS) to classify a peak as an enhancer. Default: 5000. If NULL, no filtering will be performed - use this option if your peaks are already filtered. \cr
Note that in order to filter peaks that are too close to TSS, the current \code{misha} genome must have an intervals set called \code{intervs.global.tss}.}

\item{bin_start}{the start of the trajectory. Default: 1}

\item{bin_end}{the end of the trajectory. Default: the last bin}

\item{min_initial_energy_cor}{minimal correlation between the motif normalized energy and the ATAC difference.}

\item{normalize_energies}{whether to normalize the motif energies. Set this to FALSE if the motif energies are already normalized.}

\item{energy_norm_quantile}{quantile of the energy used for normalization. Default: 1}

\item{n_prego_motifs}{number of prego motifs to consider.}

\item{traj_prego}{output of \code{learn_traj_prego}. If provided, no additional prego models would be inferred.}

\item{min_diff}{minimal ATAC difference for a peak to participate in the initial prego motif inference.}

\item{prego_sample_fraction}{Fraction of peaks to sample for prego motif inference. A smaller number would be faster but might lead to over-fitting. Default: 0.1}

\item{seed}{random seed for reproducibility.}

\item{feature_selection_beta}{beta parameter used for feature selection.}

\item{lambda}{A user supplied \code{lambda} sequence. Typical usage is to
have the program compute its own \code{lambda} sequence based on
\code{nlambda} and \code{lambda.min.ratio}. Supplying a value of
\code{lambda} overrides this. WARNING: use with care. Avoid supplying a
single value for \code{lambda} (for predictions after CV use
\code{predict()} instead).  Supply instead a decreasing sequence of
\code{lambda} values. \code{glmnet} relies on its warms starts for speed,
and its often faster to fit a whole path than compute a single fit.}

\item{alpha}{The elasticnet mixing parameter, with \eqn{0\le\alpha\le 1}.
The penalty is defined as
\deqn{(1-\alpha)/2||\beta||_2^2+\alpha||\beta||_1.} \code{alpha=1} is the
lasso penalty, and \code{alpha=0} the ridge penalty.}

\item{filter_using_r2}{whether to filter features using R^2.}

\item{r2_threshold}{minimal R^2 for a feature to be included in the model.}

\item{parallel}{whether to use parallel processing on glmnet.}

\item{peaks_size}{size of the peaks to extract sequences from. Default: 300bp}

\item{spat_num_bins}{number of spatial bins to use.}

\item{spat_bin_size}{size of each spatial bin.}

\item{kmer_sequence_length}{length of the kmer sequence to use for kmer screening. By default the full sequence is used.}
}
\value{
An instance of \code{TrajectoryModel} containing:
\itemize{
\item{model}{The final General Linear Model (GLM) object.}
\item{motif_models}{Named List, PSSM and spatial models for each motif cluster.}
\item{normalized_energies}{Numeric vector, normalized energies of each motif in each peak.}
\item{additional_features}{data frame of the additional features.}
\item{diff_score}{Numeric, normalized score of differential accessibility between 'bin_start' and 'bin_end'.}
\item{predicted_diff_score}{Numeric, predicted differential accessibility score between 'bin_start' and 'bin_end'.}
\item{initial_prego_models}{List, inferred prego models at the initial step of the algorithm.}
\item{peak_intervals}{data frame, indicating the genomic positions ('chrom', 'start', 'end') of each peak used for training.}
}
}
\description{
This function performs motif regression on ATAC trajectories. Given ATAC scores on trajectory bins, it predicts the differential accessibility between the start and end of the trajectory and returns the motifs that are most likely to be responsible for this differential accessibility.
}
