% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj-model-utils.R
\name{relearn_traj_model}
\alias{relearn_traj_model}
\title{Relearn Trajectory Model}
\usage{
relearn_traj_model(
  traj_model,
  new_energies = FALSE,
  new_logist = FALSE,
  lambda = NULL,
  use_additional_features = TRUE,
  use_motifs = TRUE,
  verbose = FALSE,
  rescale_pred = TRUE,
  relearn_model = TRUE,
  family = "binomial",
  new_interactions = FALSE,
  max_n_interactions = NULL,
  use_cv = FALSE,
  nfolds = 10,
  interaction_scale_factor = 1,
  logist_interactions = FALSE,
  logist_dinucs = FALSE,
  new_score = NULL
)
}
\arguments{
\item{traj_model}{The trajectory model object.}

\item{new_energies}{If TRUE - recreate the energies. Default is FALSE.}

\item{new_logist}{If TRUE - recreate the logistic features. Default is FALSE. If \code{new_energies} is TRUE, this is automatically set to TRUE.}

\item{lambda}{The lambda value to use for relearning. If NULL, the lambda value from the trajectory model is used.}

\item{use_additional_features}{A logical value indicating whether to use additional features. Default is TRUE.}

\item{use_motifs}{A logical value indicating whether to use motif models. Default is TRUE.}

\item{verbose}{Logical indicating whether to display additional information.}

\item{rescale_pred}{Logical indicating whether to rescale the predicted values. Default is TRUE.}

\item{relearn_model}{Logical indicating whether to relearn the model. Default is TRUE.}

\item{family}{The family to use for the glmnet model. Either "binomial" (default) or "gaussian" for linear regression.}

\item{new_interactions}{Logical indicating whether to add new interactions. Default is FALSE.}

\item{max_n_interactions}{The maximum number of interactions to add. Default is NULL.}

\item{use_cv}{Logical indicating whether to use cross-validation for lambda selection. Default is FALSE.}

\item{nfolds}{Number of folds for cross-validation. Default is 10.}

\item{interaction_scale_factor}{The factor to scale the interactions by. Default is 1 (no scaling).}

\item{logist_interactions}{Logical indicating whether to transform interactions to logistic functions. Default is FALSE.}

\item{logist_dinucs}{Logical indicating whether to transform dinucleotides to logistic functions. Default is FALSE.}

\item{new_score}{A vector of new scores to use for the model. If NULL, the diff_score is used. Default is NULL. If the vector is named, the names should be the same as the peak names in the trajectory model.}
}
\value{
The updated trajectory model object.
}
\description{
This function relearns a trajectory model using the glmnet package, without re-inferring the motif models.
}
