% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-utils.R
\name{rescale}
\alias{rescale}
\title{Rescale numeric values to a specified range}
\usage{
rescale(x, new_min = -1, new_max = 1)
}
\arguments{
\item{x}{A numeric vector of values to be rescaled.}

\item{new_min}{The minimum value of the desired output range. Defaults to -1.}

\item{new_max}{The maximum value of the desired output range. Defaults to 1.}
}
\value{
A numeric vector of rescaled values.
}
\description{
This function rescales numeric values from their original range to a specified
new range, which defaults to between -1 and 1.
}
\examples{
values <- c(2, 3, 5, 6, 8, 10)
rescale(values)
rescale(values, 0, 100)
rescale(values, 5, 50)

}
